[compiler] added a new header to center compiler definitions
