+++ src/irep2/irep2_expr.cpp
@@ -428,3 +428,36 @@ const expr2tc &object_descriptor2t::get_root_object() const
       return *tmp;
   } while(1);
 }
+
+arith_2ops::arith_2ops(
+  const type2tc &t,
+  arith_ops::expr_ids id,
+  const expr2tc &v1,
+  const expr2tc &v2)
+  : arith_ops(t, id), side_1(v1), side_2(v2)
+{
+#ifndef NDEBUG /* only check consistency in non-Release builds */
+  bool p1 = is_pointer_type(v1);
+  bool p2 = is_pointer_type(v2);
+  auto is_bv_type = [](const type2tc &t) {
+    return t->type_id == type2t::unsignedbv_id ||
+           t->type_id == type2t::signedbv_id;
+  };
+  if(p1 && p2)
+  {
+    assert(id == expr2t::sub_id);
+    assert(is_bv_type(t));
+    assert(t->get_width() == config.ansi_c.pointer_width);
+  }
+  else if(!(is_vector_type(v1->type) || is_vector_type(v2->type)))
+  {
+    assert(
+      p2 || (is_bv_type(t) == is_bv_type(v1->type) &&
+             t->get_width() == v1->type->get_width()));
+    assert(
+      p1 || (is_bv_type(t) == is_bv_type(v2->type) &&
+             t->get_width() == v2->type->get_width()));
+  }
+  // TODO: Add consistency checks for vectors
+#endif
+}
